backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
redis>=5.0.0

# Data Sources
praw>=7.7.0
google-api-python-client>=2.100.0
spotipy>=2.23.0
pytrends>=4.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Job Scheduling
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development
pytest>=7.4.0
black>=23.11.0
flake8>=6.1.0